+++ lib/timetable/timetable_screen.dart
@@ -1,5 +1,6 @@
 import 'package:buss_timetable/model/timetable.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 void navigateToTimetableRoute(BuildContext context) {
   Navigator.push(
@@ -36,47 +37,12 @@ class TimetableRoute extends StatelessWidget {
   }
 }
 
-class _TimetableSection extends StatelessWidget {
-  Timetable data;
-
-  _TimetableSection(this.data);
+class _TopAppBar extends ConsumerWidget {
+  _TopAppBar();
 
   @override
-  Widget build(BuildContext context) {
-    return Table(
-      border: TableBorder.all(width: 1.0, color: Colors.black), // 枠線の設定
-      columnWidths: const {
-        0: FlexColumnWidth(1), // 列の幅の設定
-        1: FlexColumnWidth(5),
-      },
-      children: _generateRow(data),
-    );
+  Widget build(BuildContext context, WidgetRef ref) {
+    // TODO: implement build
+    throw UnimplementedError();
   }
-}
-
-List<TableRow> _generateRow(Timetable timetable) {
-  return timetable.rows
-      .map(
-        (row) => TableRow(
-          children: [
-            Container(
-              alignment: Alignment.center,
-              padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-              color: Colors.blueGrey.shade50,
-              child: Text(row.hour.toString()),
-            ),
-            Row(
-              children: row.minute
-                  .map(
-                    (item) => Padding(
-                      padding: const EdgeInsets.only(left: 10, top: 4),
-                      child: Text(item.toString().padLeft(2, '0')),
-                    ),
-                  )
-                  .toList(),
-            )
-          ],
-        ),
-      )
-      .toList();
 }
